refactor Header component
