+++ src/components/Header/Button.js
@@ -1,7 +1,7 @@
 import { React } from 'react'
 import { css } from '@emotion/core'
 import { Link } from 'gatsby'
-import { useTheme } from './Theming'
+import { useTheme } from '../Theming'
 import { rgba, darken } from 'polished'
 
 const Button = ({ to, children, secondary, ...restProps }) => {
