+++ src/components/Header/Links.js
@@ -1,0 +1,28 @@
+import React from 'react'
+import { Link, StaticQuery, graphql } from 'gatsby'
+import { css } from '@emotion/core'
+import { useTheme } from '../Theming'
+import ThemeToggler from './ThemeToggler'
+
+export default () => {
+  const theme = useTheme()
+  return (
+    <React.Fragment>
+      <Link to="#" activeClassName="active" aria-label="View blog page">
+        Blog
+      </Link>
+      <Link to="#" activeClassName="active" aria-label="View blog page">
+        About
+      </Link>
+      <Link to="#" activeClassName="active" aria-label="View blog page">
+        Contact
+      </Link>
+
+      <ThemeToggler
+        css={{}}
+        toggleTheme={theme.toggleTheme}
+        themeName={theme.themeName}
+      />
+    </React.Fragment>
+  )
+}
